# chore: migrate to ts 5.2.2