--- a/apps/react-18-tests-v8/tsconfig.cy.json
+++ b/apps/react-18-tests-v8/tsconfig.cy.json
@@ -4,6 +4,7 @@
     "jsx": "react",
     "isolatedModules": false,
     "types": ["node", "cypress", "cypress-real-events"],
+    "typeRoots": ["../../node_modules", "../../node_modules/@types"],
     "lib": ["ES2019", "dom"]
   },
   "include": ["./src/**/*.cy.ts", "./src/**/*.cy.tsx"]