--- a/apps/vr-tests-web-components/tsconfig.json
+++ b/apps/vr-tests-web-components/tsconfig.json
@@ -2,9 +2,10 @@
   "extends": "../../tsconfig.base.wc.json",
   "compilerOptions": {
     "target": "ES2019",
-    "module": "ESNext",
+    "module": "NodeNext",
     "noEmit": true,
     "experimentalDecorators": true,
+    "moduleResolution": "NodeNext",
     "resolveJsonModule": true,
     "allowJs": true,
     "jsx": "react",