--- a/change/@fluentui-babel-preset-global-context-28602215-c268-452d-85d9-00ba63d8ed0f.json
+++ b/change/@fluentui-babel-preset-global-context-28602215-c268-452d-85d9-00ba63d8ed0f.json
@@ -1,0 +1,7 @@
+{
+  "type": "none",
+  "comment": "added explicit typeRoot to tsconfig.json",
+  "packageName": "@fluentui/babel-preset-global-context",
+  "email": "viktorgenaev@microsoft.com",
+  "dependentChangeType": "none"
+}