--- a/change/@fluentui-eslint-plugin-d576152d-5ad5-451e-8384-e81fd4d2bb98.json
+++ b/change/@fluentui-eslint-plugin-d576152d-5ad5-451e-8384-e81fd4d2bb98.json
@@ -1,0 +1,7 @@
+{
+  "type": "none",
+  "comment": "update tsconfig after migration to 5.2",
+  "packageName": "@fluentui/eslint-plugin",
+  "email": "viktorgenaev@microsoft.com",
+  "dependentChangeType": "none"
+}