--- a/change/@fluentui-global-context-85d11653-4f77-49c6-83d5-ebe3dc5c969e.json
+++ b/change/@fluentui-global-context-85d11653-4f77-49c6-83d5-ebe3dc5c969e.json
@@ -1,0 +1,7 @@
+{
+  "type": "none",
+  "comment": "add explicit typeRoot to tsconfig.json",
+  "packageName": "@fluentui/global-context",
+  "email": "viktorgenaev@microsoft.com",
+  "dependentChangeType": "none"
+}