--- a/change/@fluentui-react-aria-fd4f3898-30c0-423a-ad87-c68621431b95.json
+++ b/change/@fluentui-react-aria-fd4f3898-30c0-423a-ad87-c68621431b95.json
@@ -1,0 +1,7 @@
+{
+  "type": "none",
+  "comment": "add explicit typeRoot to tsconfig.json",
+  "packageName": "@fluentui/react-aria",
+  "email": "viktorgenaev@microsoft.com",
+  "dependentChangeType": "none"
+}