--- a/change/@fluentui-react-avatar-1f8e1036-8cb1-4b96-9155-16a930f92f54.json
+++ b/change/@fluentui-react-avatar-1f8e1036-8cb1-4b96-9155-16a930f92f54.json
@@ -1,0 +1,7 @@
+{
+  "type": "none",
+  "comment": "add explicit typeRoot to tsconfig.json",
+  "packageName": "@fluentui/react-avatar",
+  "email": "viktorgenaev@microsoft.com",
+  "dependentChangeType": "none"
+}