--- a/change/@fluentui-react-breadcrumb-517d614c-0eb1-4019-aa94-a8011d3947a6.json
+++ b/change/@fluentui-react-breadcrumb-517d614c-0eb1-4019-aa94-a8011d3947a6.json
@@ -1,0 +1,7 @@
+{
+  "type": "none",
+  "comment": "add explicit typeRoot to tsconfig.json",
+  "packageName": "@fluentui/react-breadcrumb",
+  "email": "viktorgenaev@microsoft.com",
+  "dependentChangeType": "none"
+}