--- a/change/@fluentui-react-card-c6d87bf2-dcd8-488e-a2d2-b57ea74cc51d.json
+++ b/change/@fluentui-react-card-c6d87bf2-dcd8-488e-a2d2-b57ea74cc51d.json
@@ -1,0 +1,7 @@
+{
+  "type": "none",
+  "comment": "add explicit typeRoot to tsconfig.json",
+  "packageName": "@fluentui/react-card",
+  "email": "viktorgenaev@microsoft.com",
+  "dependentChangeType": "none"
+}