--- a/change/@fluentui-react-combobox-fe9850c4-ca14-4c22-a7c8-a869c54ad106.json
+++ b/change/@fluentui-react-combobox-fe9850c4-ca14-4c22-a7c8-a869c54ad106.json
@@ -1,0 +1,7 @@
+{
+  "type": "none",
+  "comment": "add explicit typeRoot to tsconfig.json",
+  "packageName": "@fluentui/react-combobox",
+  "email": "viktorgenaev@microsoft.com",
+  "dependentChangeType": "none"
+}