--- a/change/@fluentui-react-datepicker-compat-a54d07d4-b0fd-4b75-8a3e-3d9851d91133.json
+++ b/change/@fluentui-react-datepicker-compat-a54d07d4-b0fd-4b75-8a3e-3d9851d91133.json
@@ -1,0 +1,7 @@
+{
+  "type": "none",
+  "comment": "add explicit typeRoot to tsconfig.json",
+  "packageName": "@fluentui/react-datepicker-compat",
+  "email": "viktorgenaev@microsoft.com",
+  "dependentChangeType": "none"
+}