--- a/change/@fluentui-react-dialog-fc005c16-f756-4b47-b5e3-f544abbae52c.json
+++ b/change/@fluentui-react-dialog-fc005c16-f756-4b47-b5e3-f544abbae52c.json
@@ -1,0 +1,7 @@
+{
+  "type": "none",
+  "comment": "add explicit typeRoot to tsconfig.json",
+  "packageName": "@fluentui/react-dialog",
+  "email": "viktorgenaev@microsoft.com",
+  "dependentChangeType": "none"
+}