--- a/change/@fluentui-react-drawer-edb16867-7903-49df-9445-e5d40af61f73.json
+++ b/change/@fluentui-react-drawer-edb16867-7903-49df-9445-e5d40af61f73.json
@@ -1,0 +1,7 @@
+{
+  "type": "none",
+  "comment": "add explicit typeRoot to tsconfig.json",
+  "packageName": "@fluentui/react-drawer",
+  "email": "viktorgenaev@microsoft.com",
+  "dependentChangeType": "none"
+}