--- a/change/@fluentui-react-infolabel-dc7c6223-4a06-42f1-a6f6-9ddaf488803e.json
+++ b/change/@fluentui-react-infolabel-dc7c6223-4a06-42f1-a6f6-9ddaf488803e.json
@@ -1,0 +1,7 @@
+{
+  "type": "none",
+  "comment": "add explicit typeRoot to tsconfig.json",
+  "packageName": "@fluentui/react-infolabel",
+  "email": "viktorgenaev@microsoft.com",
+  "dependentChangeType": "none"
+}