--- a/change/@fluentui-react-list-preview-f355ed2f-294e-4b0d-8f0c-59f684d61515.json
+++ b/change/@fluentui-react-list-preview-f355ed2f-294e-4b0d-8f0c-59f684d61515.json
@@ -1,0 +1,7 @@
+{
+  "type": "none",
+  "comment": "add explicit typeRoot to tsconfig.json",
+  "packageName": "@fluentui/react-list-preview",
+  "email": "viktorgenaev@microsoft.com",
+  "dependentChangeType": "none"
+}