--- a/change/@fluentui-react-menu-19683f73-d2c1-441e-87fa-f4ca074e7bd2.json
+++ b/change/@fluentui-react-menu-19683f73-d2c1-441e-87fa-f4ca074e7bd2.json
@@ -1,0 +1,7 @@
+{
+  "type": "none",
+  "comment": "add explicit typeRoot to tsconfig.json",
+  "packageName": "@fluentui/react-menu",
+  "email": "viktorgenaev@microsoft.com",
+  "dependentChangeType": "none"
+}