--- a/change/@fluentui-react-migration-v0-v9-31f1ddf5-6ce6-424a-a576-8fb4b8336d63.json
+++ b/change/@fluentui-react-migration-v0-v9-31f1ddf5-6ce6-424a-a576-8fb4b8336d63.json
@@ -1,0 +1,7 @@
+{
+  "type": "none",
+  "comment": "update generated api",
+  "packageName": "@fluentui/react-migration-v0-v9",
+  "email": "viktorgenaev@microsoft.com",
+  "dependentChangeType": "none"
+}