--- a/change/@fluentui-react-motion-1ecb79e8-deb8-4e5e-8be4-341fcbd70ead.json
+++ b/change/@fluentui-react-motion-1ecb79e8-deb8-4e5e-8be4-341fcbd70ead.json
@@ -1,0 +1,7 @@
+{
+  "type": "none",
+  "comment": "add explicit typeRoot to tsconfig.json",
+  "packageName": "@fluentui/react-motion",
+  "email": "viktorgenaev@microsoft.com",
+  "dependentChangeType": "none"
+}