--- a/change/@fluentui-react-overflow-88fb1633-a8a7-4c6e-827c-f48b97702a3f.json
+++ b/change/@fluentui-react-overflow-88fb1633-a8a7-4c6e-827c-f48b97702a3f.json
@@ -1,0 +1,7 @@
+{
+  "type": "none",
+  "comment": "add explicit typeRoot to tsconfig.json",
+  "packageName": "@fluentui/react-overflow",
+  "email": "viktorgenaev@microsoft.com",
+  "dependentChangeType": "none"
+}