--- a/change/@fluentui-react-popover-23093573-92f3-4ac1-9032-894182562d8e.json
+++ b/change/@fluentui-react-popover-23093573-92f3-4ac1-9032-894182562d8e.json
@@ -1,0 +1,7 @@
+{
+  "type": "none",
+  "comment": "add explicit typeRoot to tsconfig.json",
+  "packageName": "@fluentui/react-popover",
+  "email": "viktorgenaev@microsoft.com",
+  "dependentChangeType": "none"
+}