--- a/change/@fluentui-react-portal-compat-fac2b4d7-61f4-4bcb-932e-051821f42377.json
+++ b/change/@fluentui-react-portal-compat-fac2b4d7-61f4-4bcb-932e-051821f42377.json
@@ -1,0 +1,7 @@
+{
+  "type": "none",
+  "comment": "add explicit typeRoot to tsconfig.json",
+  "packageName": "@fluentui/react-portal-compat",
+  "email": "viktorgenaev@microsoft.com",
+  "dependentChangeType": "none"
+}