--- a/change/@fluentui-react-swatch-picker-3866ae6e-34e1-45c7-b9a5-66333c6f2e4e.json
+++ b/change/@fluentui-react-swatch-picker-3866ae6e-34e1-45c7-b9a5-66333c6f2e4e.json
@@ -1,0 +1,7 @@
+{
+  "type": "none",
+  "comment": "add explicit typeRoot to tsconfig.json",
+  "packageName": "@fluentui/react-swatch-picker",
+  "email": "viktorgenaev@microsoft.com",
+  "dependentChangeType": "none"
+}