--- a/change/@fluentui-react-table-dfd62e19-0fab-4f85-8713-15df7f8770cb.json
+++ b/change/@fluentui-react-table-dfd62e19-0fab-4f85-8713-15df7f8770cb.json
@@ -1,0 +1,7 @@
+{
+  "type": "none",
+  "comment": "add explicit typeRoot to tsconfig.json",
+  "packageName": "@fluentui/react-table",
+  "email": "viktorgenaev@microsoft.com",
+  "dependentChangeType": "none"
+}