--- a/change/@fluentui-react-tabster-73a95c93-994c-40ad-8a20-e6b93ce230cc.json
+++ b/change/@fluentui-react-tabster-73a95c93-994c-40ad-8a20-e6b93ce230cc.json
@@ -1,0 +1,7 @@
+{
+  "type": "none",
+  "comment": "add explicit typeRoot to tsconfig.json",
+  "packageName": "@fluentui/react-tabster",
+  "email": "viktorgenaev@microsoft.com",
+  "dependentChangeType": "none"
+}