--- a/change/@fluentui-react-tag-picker-c37fc1e6-1a66-491a-9c29-03eb496caeea.json
+++ b/change/@fluentui-react-tag-picker-c37fc1e6-1a66-491a-9c29-03eb496caeea.json
@@ -1,0 +1,7 @@
+{
+  "type": "none",
+  "comment": "add explicit typeRoot to tsconfig.json",
+  "packageName": "@fluentui/react-tag-picker",
+  "email": "viktorgenaev@microsoft.com",
+  "dependentChangeType": "none"
+}