--- a/change/@fluentui-react-tags-91f4933e-8018-4cea-8440-417bbaf4e7d6.json
+++ b/change/@fluentui-react-tags-91f4933e-8018-4cea-8440-417bbaf4e7d6.json
@@ -1,0 +1,7 @@
+{
+  "type": "none",
+  "comment": "add explicit typeRoot to tsconfig.json",
+  "packageName": "@fluentui/react-tags",
+  "email": "viktorgenaev@microsoft.com",
+  "dependentChangeType": "none"
+}