--- a/change/@fluentui-react-timepicker-compat-d4c11b3d-1b54-44e7-8231-c31f5e2eb183.json
+++ b/change/@fluentui-react-timepicker-compat-d4c11b3d-1b54-44e7-8231-c31f5e2eb183.json
@@ -1,0 +1,7 @@
+{
+  "type": "none",
+  "comment": "add explicit typeRoot to tsconfig.json",
+  "packageName": "@fluentui/react-timepicker-compat",
+  "email": "viktorgenaev@microsoft.com",
+  "dependentChangeType": "none"
+}