--- a/change/@fluentui-react-toast-868f2a7b-e5f8-4fde-8890-c2d679213403.json
+++ b/change/@fluentui-react-toast-868f2a7b-e5f8-4fde-8890-c2d679213403.json
@@ -1,0 +1,7 @@
+{
+  "type": "none",
+  "comment": "add explicit typeRoot to tsconfig.json",
+  "packageName": "@fluentui/react-toast",
+  "email": "viktorgenaev@microsoft.com",
+  "dependentChangeType": "none"
+}