--- a/change/@fluentui-react-toolbar-5bd96504-b775-4b48-b161-765a58dab3a4.json
+++ b/change/@fluentui-react-toolbar-5bd96504-b775-4b48-b161-765a58dab3a4.json
@@ -1,0 +1,7 @@
+{
+  "type": "none",
+  "comment": "add explicit typeRoot to tsconfig.json",
+  "packageName": "@fluentui/react-toolbar",
+  "email": "viktorgenaev@microsoft.com",
+  "dependentChangeType": "none"
+}