--- a/change/@fluentui-react-tree-4b796704-0c2a-4529-89fc-7f66f1de0e14.json
+++ b/change/@fluentui-react-tree-4b796704-0c2a-4529-89fc-7f66f1de0e14.json
@@ -1,0 +1,7 @@
+{
+  "type": "none",
+  "comment": "add explicit typeRoot to tsconfig.json",
+  "packageName": "@fluentui/react-tree",
+  "email": "viktorgenaev@microsoft.com",
+  "dependentChangeType": "none"
+}