--- a/change/@fluentui-react-utilities-a76b7036-25c3-49fa-bf0a-8cab44d8d4eb.json
+++ b/change/@fluentui-react-utilities-a76b7036-25c3-49fa-bf0a-8cab44d8d4eb.json
@@ -1,0 +1,7 @@
+{
+  "type": "none",
+  "comment": "add explicit typeRoot to tsconfig.json",
+  "packageName": "@fluentui/react-utilities",
+  "email": "viktorgenaev@microsoft.com",
+  "dependentChangeType": "none"
+}