--- a/change/@fluentui-theme-18bd688f-f330-4fb2-abc6-d70673fe6bc6.json
+++ b/change/@fluentui-theme-18bd688f-f330-4fb2-abc6-d70673fe6bc6.json
@@ -1,0 +1,7 @@
+{
+  "type": "none",
+  "comment": "update api.md",
+  "packageName": "@fluentui/theme",
+  "email": "viktorgenaev@microsoft.com",
+  "dependentChangeType": "none"
+}