--- a/change/@fluentui-web-components-a0546391-be46-4cec-ba6d-1b268e7f08b0.json
+++ b/change/@fluentui-web-components-a0546391-be46-4cec-ba6d-1b268e7f08b0.json
@@ -1,0 +1,7 @@
+{
+  "type": "none",
+  "comment": "update tsconfig to be compatiable with 5.2",
+  "packageName": "@fluentui/web-components",
+  "email": "viktorgenaev@microsoft.com",
+  "dependentChangeType": "none"
+}