--- a/packages/react-components/react-tree/library/tsconfig.cy.json
+++ b/packages/react-components/react-tree/library/tsconfig.cy.json
@@ -2,7 +2,8 @@
   "extends": "./tsconfig.json",
   "compilerOptions": {
     "isolatedModules": false,
-    "types": ["node", "cypress", "cypress-real-events"]
+    "types": ["node", "cypress", "cypress-real-events"],
+    "typeRoots": ["../../../../node_modules", "../../../../node_modules/@types"]
   },
   "include": ["**/*.cy.ts", "**/*.cy.tsx"]
 }