--- a/packages/react-examples/tsconfig.cy.json
+++ b/packages/react-examples/tsconfig.cy.json
@@ -3,6 +3,7 @@
   "compilerOptions": {
     "isolatedModules": false,
     "types": ["node", "cypress", "cypress-real-events"],
+    "typeRoots": ["../../../node_modules", "../../../node_modules/@types"],
     "lib": ["ES2019", "dom"]
   },
   "include": ["src/**/*.cy.ts", "src/**/*.cy.tsx"]