--- a/packages/web-components/src/menu-item/menu-item.template.ts
+++ b/packages/web-components/src/menu-item/menu-item.template.ts
@@ -18,9 +18,7 @@ export function menuItemTemplate<T extends MenuItem>(options: MenuItemOptions = 
       @click="${(x, c) => x.handleMenuItemClick(c.event as MouseEvent)}"
       @mouseover="${(x, c) => x.handleMouseOver(c.event as MouseEvent)}"
       @mouseout="${(x, c) => x.handleMouseOut(c.event as MouseEvent)}"
-      @toggle="${(x, c) =>
-        // @ts-expect-error - Baseline 2024
-        x.toggleHandler(c.event as ToggleEvent)}"
+      @toggle="${(x, c) => x.toggleHandler(c.event as ToggleEvent)}"
     >
       <slot name="indicator"> ${staticallyCompose(options.indicator)} </slot>
       ${startSlotTemplate(options)}