--- a/packages/web-components/tsconfig.json
+++ b/packages/web-components/tsconfig.json
@@ -2,8 +2,8 @@
   "extends": "../../tsconfig.base.wc.json",
   "compilerOptions": {
     "target": "ES2019",
-    "module": "NodeNext",
-    "moduleResolution": "Node16",
+    "module": "ESNext",
+    "moduleResolution": "bundler",
     "experimentalDecorators": true,
     "resolveJsonModule": true,
     "allowJs": true