--- a/tools/workspace-plugin/src/generators/cypress-component-configuration/files/tsconfig.cy.json__tmpl__
+++ b/tools/workspace-plugin/src/generators/cypress-component-configuration/files/tsconfig.cy.json__tmpl__
@@ -3,6 +3,7 @@
   "compilerOptions":{
     "isolatedModules":false,
     "types":["node", "cypress", "cypress-real-events"],
+    "typeRoots":["<%= options.rootOffset %>node_modules", "<%= options.rootOffset %>node_modules/@types"],
     "lib":["ES2019", "dom"]
   },
   "include":["**/*.cy.ts", "**/*.cy.tsx"]