… packages/eslint-plugin/tsconfig.json
… packages/web-components/tsconfig.lib.json
… packages/web-components/tsconfig.spec.json
… tools/eslint-rules/tsconfig.json
… tsconfig.base.wc.json
… yarn.lock