<!--
Thank you for submitting a pull request!

Please verify that:
* [ ] Code is up-to-date with the `master` branch
* [ ] Your changes are covered by tests (if possible)
* [ ] You've run `yarn change` locally


PR flow tips:
* [ ] Try to start with a Draft PR
* [ ] Once you're ready (ideally the pipeline is passing) promote your PR to Ready for Review. This step will auto-assign reviewers for your PR.
-->

Release notes: 
- 5.1: https://www.typescriptlang.org/docs/handbook/release-notes/typescript-5-1.html
- 5.2: https://www.typescriptlang.org/docs/handbook/release-notes/typescript-5-2.html 

What's addressed: 
- Explicit type roots. [Type roots are no longer making upward walking.](https://www.typescriptlang.org/docs/handbook/release-notes/typescript-5-1.html#explicit-typeroots-disables-upward-walks-for-node_modulestypes). It's not ideal and related to current tsconfig nature, files are relative to the location of the file itself. We will fix this in version 5.5, where [`${configDir}` is introduced.](https://devblogs.microsoft.com/typescript/announcing-typescript-5-5/#the-$%7Bconfigdir%7D-template-variable-for-configuration-files)
- Module and ModuleResolution since 5.2 [should match recent node settigns](https://www.typescriptlang.org/docs/handbook/release-notes/typescript-5-2.html#module-and-moduleresolution-must-match-under-recent-nodejs-settings)

## Previous Behavior

<!-- This is the behavior we have today -->

## New Behavior

<!-- This is the behavior we should expect with the changes in this PR -->

## Related Issue(s)

<!-- Please link the issue being fixed so it gets closed when this is merged. -->

A little slowdown should be caused by explicit type roots, once we will update to ts 5.5 -> it should get back to normal or even faster type-check. 

| Before (5.2) | After (5.2) |
|--------|--------|
| 7m 7s | 7m 20s / **𝚫 3% 🐢  SLOWER** | 

- Implements #32402